Refacor code
